[DF-14522] Review fixes
Former-commit-id: 4b693c5dd9c8f610e23f16459a9b884602b5df7f

+++ Sources/Internal/Base/FastName.cpp
@@ -54,7 +54,7 @@ void FastName::Init(const char* name)
         db->sizeOfNames += (nameLen * sizeof(FastNameDB::CharT));
 
         // index will be a new row in names table
-        size_t index = static_cast<int32>(db->namesTable.size());
+        size_t index = db->namesTable.size();
 
         db->namesTable.push_back(nameCopy);
         db->nameToIndexMap.emplace(nameCopy, index);
